move queuing PR comments task into CommitContextIntegration ABC
